requirements.txt
fastapi
uvicorn
opencv-python
numpy
python-dotenv
requests
cloudinary
python-multipart
